+++ app/src/main/java/com/heyfood/heyfoodapp/contato/persistencia/ContatoDAO.java
@@ -19,11 +19,11 @@ public class ContatoDAO extends AbstractDAO{
         helper = new DBHelper(context);
     }
 
-    public Contato getContato(int id) {
+    public Contato getContato(long id) {
         Contato result = null;
         db = helper.getReadableDatabase();
         String sql = "SELECT * FROM " + DBHelper.TABELA_CONTATO+ " WHERE " + DBHelper.CAMPO_ID_CONTATO + " LIKE ?;";
-        Cursor cursor = db.rawQuery(sql, new String[]{Integer.toString(id)});
+        Cursor cursor = db.rawQuery(sql, new String[]{Long.toString(id)});
         if (cursor.moveToFirst()) {
             result = createContato(cursor);
         }
